+++ features/suite_test.go
@@ -703,6 +703,7 @@ func initializeScenario(sc *godog.ScenarioContext) {
 	initializeDispatchingSteps(sc)
 	initializeWaitsSteps(sc)
 	initializeHeadroomSteps(sc)
+	initializeAdmissionSteps(sc)
 	initializeWorkingSteps(sc)
 	initializeDrainSteps(sc)
 	initializeHookSteps(sc)